Work around crashes to demand import not finding _speedups.
The speedups.c module is a C extension that is not emmbedded in this copy of simplejson. For some reasons, when demandimport is enabled, it crashes failing to localize accelerated C function.
As we do not need highly accelerated C version of simplejson, mark speedups function as absent so they are not queried.