disable proxy detection, added --longdiff option
i) We wanted to specifically disable the proxy auto detection, as urlib2
does not seem to obey/follow the proxy exclusions(registered in windows at
least.)
   --- Iv'e added an 'http_proxy' control variable to the config to provide
this

ii) We wanted to configure or detect the right repository id
   --- Iv'e added a 'repo_id' config variable to control this, we will set
this in the project .hg/hgrc file.

iii) We want to automatically pick the common ancestor as the review parent
   --- I have copied your base diff methods and provided a --longdiff
option to auto magically find the ancestral parent for review.